Update file structure & Update parameters
* Update file structure
* Delete unnecessary getters
